+++ core/src/components/modal/modal-interface.ts
@@ -25,6 +25,7 @@ export interface ModalOptions<T extends ComponentRef = ComponentRef> {
   backdropBreakpoint?: number;
   handle?: boolean;
   handleBehavior?: ModalHandleBehavior;
+  expandToScroll?: boolean;
 }
 
 export interface ModalAnimationOptions {