+++ core/src/components/select/test/highlight/select.e2e.ts
@@ -7,41 +7,114 @@ configs({ modes: ['md'], directions: ['ltr'] }).forEach(({ title, screenshot, co
       test('should render valid state correctly', async ({ page }) => {
         await page.setContent(
           `
-          <ion-select value="apple" class="ion-valid ion-focused" label="Favorite Fruit">
-            <ion-select-option value="apple">Apple</ion-select-option>
-          </ion-select>
+          <style>
+            #container {
+              display: grid;
+              gap: 5px;
+            }
+          </style>
+          <div id="container">
+            <ion-select value="apple" class="ion-valid has-focus" label="Favorite Fruit">
+              <ion-select-option value="apple">Apple</ion-select-option>
+            </ion-select>
+            <ion-list>
+              <ion-item>
+                <ion-select value="apple" class="ion-valid has-focus" label="Favorite Fruit">
+                  <ion-select-option value="apple">Apple</ion-select-option>
+                </ion-select>
+              </ion-item>
+            </ion-list>
+          </div>
         `,
           config
         );
 
-        const select = page.locator('ion-select');
-        await expect(select).toHaveScreenshot(screenshot(`select-no-fill-valid`));
+        const container = page.locator('#container');
+        await expect(container).toHaveScreenshot(screenshot(`select-no-fill-valid`));
       });
       test('should render invalid state correctly', async ({ page }) => {
         await page.setContent(
           `
-          <ion-select value="apple" class="ion-touched ion-invalid" label="Favorite Fruit">
-            <ion-select-option value="apple">Apple</ion-select-option>
-          </ion-select>
+          <style>
+            #container {
+              display: grid;
+              gap: 5px;
+            }
+          </style>
+          <div id="container">
+            <ion-select value="apple" class="ion-touched ion-invalid" label="Favorite Fruit">
+              <ion-select-option value="apple">Apple</ion-select-option>
+            </ion-select>
+            <ion-list>
+              <ion-item>
+                <ion-select value="apple" class="ion-touched ion-invalid" label="Favorite Fruit">
+                  <ion-select-option value="apple">Apple</ion-select-option>
+                </ion-select>
+              </ion-item>
+            </ion-list>
+          </div>
         `,
           config
         );
 
-        const select = page.locator('ion-select');
-        await expect(select).toHaveScreenshot(screenshot(`select-no-fill-invalid`));
+        const container = page.locator('#container');
+        await expect(container).toHaveScreenshot(screenshot(`select-no-fill-invalid`));
+      });
+      test('should render invalid with focus state correctly', async ({ page }) => {
+        await page.setContent(
+          `
+          <style>
+            #container {
+              display: grid;
+              gap: 5px;
+            }
+          </style>
+          <div id="container">
+            <ion-select value="apple" class="ion-touched ion-invalid has-focus" label="Favorite Fruit">
+              <ion-select-option value="apple">Apple</ion-select-option>
+            </ion-select>
+            <ion-list>
+              <ion-item>
+                <ion-select value="apple" class="ion-touched ion-invalid has-focus" label="Favorite Fruit">
+                  <ion-select-option value="apple">Apple</ion-select-option>
+                </ion-select>
+              </ion-item>
+            </ion-list>
+          </div>
+        `,
+          config
+        );
+
+        const container = page.locator('#container');
+        await expect(container).toHaveScreenshot(screenshot(`select-no-fill-invalid-focus`));
       });
       test('should render focused state correctly', async ({ page }) => {
         await page.setContent(
           `
-          <ion-select value="apple" class="ion-focused" label="Favorite Fruit">
-            <ion-select-option value="apple">Apple</ion-select-option>
-          </ion-select>
+          <style>
+            #container {
+              display: grid;
+              gap: 5px;
+            }
+          </style>
+          <div id="container">
+            <ion-select value="apple" class="has-focus" label="Favorite Fruit">
+              <ion-select-option value="apple">Apple</ion-select-option>
+            </ion-select>
+            <ion-list>
+              <ion-item>
+                <ion-select value="apple" class="has-focus" label="Favorite Fruit">
+                  <ion-select-option value="apple">Apple</ion-select-option>
+                </ion-select>
+              </ion-item>
+            </ion-list>
+          </div>
         `,
           config
         );
 
-        const select = page.locator('ion-select');
-        await expect(select).toHaveScreenshot(screenshot(`select-no-fill-focus`));
+        const container = page.locator('#container');
+        await expect(container).toHaveScreenshot(screenshot(`select-no-fill-focus`));
       });
       test('should render custom highlight correctly', async ({ page }) => {
         await page.setContent(
@@ -56,15 +129,15 @@ configs({ modes: ['md'], directions: ['ltr'] }).forEach(({ title, screenshot, co
           </style>
 
           <div class="container">
-            <ion-select value="apple" class="custom ion-valid ion-focused" label="Favorite Fruit">
+            <ion-select value="apple" class="custom ion-valid has-focus" label="Favorite Fruit">
               <ion-select-option value="apple">Apple</ion-select-option>
             </ion-select>
 
-            <ion-select value="apple" class="custom ion-touched ion-invalid ion-focused" label="Favorite Fruit">
+            <ion-select value="apple" class="custom ion-touched ion-invalid has-focus" label="Favorite Fruit">
               <ion-select-option value="apple">Apple</ion-select-option>
             </ion-select>
 
-            <ion-select value="apple" class="custom ion-focused" label="Favorite Fruit">
+            <ion-select value="apple" class="custom has-focus" label="Favorite Fruit">
               <ion-select-option value="apple">Apple</ion-select-option>
             </ion-select>
           </div>
@@ -80,41 +153,86 @@ configs({ modes: ['md'], directions: ['ltr'] }).forEach(({ title, screenshot, co
       test('should render valid state correctly', async ({ page }) => {
         await page.setContent(
           `
-          <ion-select fill="solid" value="apple" class="ion-valid ion-focused" label="Favorite Fruit">
-            <ion-select-option value="apple">Apple</ion-select-option>
-          </ion-select>
+          <style>
+            #container {
+              display: grid;
+              gap: 5px;
+            }
+          </style>
+          <div id="container">
+            <ion-select fill="solid" value="apple" class="ion-valid has-focus" label="Favorite Fruit">
+              <ion-select-option value="apple">Apple</ion-select-option>
+            </ion-select>
+            <ion-list>
+              <ion-item>
+                <ion-select fill="solid" value="apple" class="ion-valid has-focus" label="Favorite Fruit">
+                  <ion-select-option value="apple">Apple</ion-select-option>
+                </ion-select>
+              </ion-item>
+            </ion-list>
+          </div>
         `,
           config
         );
 
-        const select = page.locator('ion-select');
-        await expect(select).toHaveScreenshot(screenshot(`select-solid-valid`));
+        const container = page.locator('#container');
+        await expect(container).toHaveScreenshot(screenshot(`select-solid-valid`));
       });
       test('should render invalid state correctly', async ({ page }) => {
         await page.setContent(
           `
-          <ion-select fill="solid" value="apple" class="ion-touched ion-invalid" label="Favorite Fruit">
-            <ion-select-option value="apple">Apple</ion-select-option>
-          </ion-select>
+          <style>
+            #container {
+              display: grid;
+              gap: 5px;
+            }
+          </style>
+          <div id="container">
+            <ion-select fill="solid" value="apple" class="ion-touched ion-invalid" label="Favorite Fruit">
+              <ion-select-option value="apple">Apple</ion-select-option>
+            </ion-select>
+            <ion-list>
+              <ion-item>
+                <ion-select fill="solid" value="apple" class="ion-touched ion-invalid" label="Favorite Fruit">
+                  <ion-select-option value="apple">Apple</ion-select-option>
+                </ion-select>
+              </ion-item>
+            </ion-list>
+          </div>
         `,
           config
         );
 
-        const select = page.locator('ion-select');
-        await expect(select).toHaveScreenshot(screenshot(`select-solid-invalid`));
+        const container = page.locator('#container');
+        await expect(container).toHaveScreenshot(screenshot(`select-solid-invalid`));
       });
       test('should render focused state correctly', async ({ page }) => {
         await page.setContent(
           `
-          <ion-select fill="solid" value="apple" class="ion-focused" label="Favorite Fruit">
-            <ion-select-option value="apple">Apple</ion-select-option>
-          </ion-select>
+          <style>
+            #container {
+              display: grid;
+              gap: 5px;
+            }
+          </style>
+          <div id="container">
+            <ion-select fill="solid" value="apple" class="has-focus" label="Favorite Fruit">
+              <ion-select-option value="apple">Apple</ion-select-option>
+            </ion-select>
+            <ion-list>
+              <ion-item>
+                <ion-select fill="solid" value="apple" class="has-focus" label="Favorite Fruit">
+                  <ion-select-option value="apple">Apple</ion-select-option>
+                </ion-select>
+              </ion-item>
+            </ion-list>
+          </div>
         `,
           config
         );
 
-        const select = page.locator('ion-select');
-        await expect(select).toHaveScreenshot(screenshot(`select-solid-focus`));
+        const container = page.locator('#container');
+        await expect(container).toHaveScreenshot(screenshot(`select-solid-focus`));
       });
       test('should render custom highlight correctly', async ({ page }) => {
         await page.setContent(
@@ -129,15 +247,15 @@ configs({ modes: ['md'], directions: ['ltr'] }).forEach(({ title, screenshot, co
           </style>
 
           <div class="container">
-            <ion-select fill="solid" value="apple" class="custom ion-valid ion-focused" label="Favorite Fruit">
+            <ion-select fill="solid" value="apple" class="custom ion-valid has-focus" label="Favorite Fruit">
               <ion-select-option value="apple">Apple</ion-select-option>
             </ion-select>
 
-            <ion-select fill="solid" value="apple" class="custom ion-touched ion-invalid ion-focused" label="Favorite Fruit">
+            <ion-select fill="solid" value="apple" class="custom ion-touched ion-invalid has-focus" label="Favorite Fruit">
               <ion-select-option value="apple">Apple</ion-select-option>
             </ion-select>
 
-            <ion-select fill="solid" value="apple" class="custom ion-focused" label="Favorite Fruit">
+            <ion-select fill="solid" value="apple" class="custom has-focus" label="Favorite Fruit">
               <ion-select-option value="apple">Apple</ion-select-option>
             </ion-select>
           </div>
@@ -153,41 +271,86 @@ configs({ modes: ['md'], directions: ['ltr'] }).forEach(({ title, screenshot, co
       test('should render valid state correctly', async ({ page }) => {
         await page.setContent(
           `
-          <ion-select fill="outline" value="apple" class="ion-valid ion-focused" label="Favorite Fruit">
-            <ion-select-option value="apple">Apple</ion-select-option>
-          </ion-select>
+          <style>
+            #container {
+              display: grid;
+              gap: 5px;
+            }
+          </style>
+          <div id="container">
+            <ion-select fill="outline" value="apple" class="ion-valid has-focus" label="Favorite Fruit">
+              <ion-select-option value="apple">Apple</ion-select-option>
+            </ion-select>
+            <ion-list>
+              <ion-item>
+                <ion-select fill="outline" value="apple" class="ion-valid has-focus" label="Favorite Fruit">
+                  <ion-select-option value="apple">Apple</ion-select-option>
+                </ion-select>
+              </ion-item>
+            </ion-list>
+          </div>
         `,
           config
         );
 
-        const select = page.locator('ion-select');
-        await expect(select).toHaveScreenshot(screenshot(`select-outline-valid`));
+        const container = page.locator('#container');
+        await expect(container).toHaveScreenshot(screenshot(`select-outline-valid`));
       });
       test('should render invalid state correctly', async ({ page }) => {
         await page.setContent(
           `
-          <ion-select fill="outline" value="apple" class="ion-touched ion-invalid ion-focused" label="Favorite Fruit">
-            <ion-select-option value="apple">Apple</ion-select-option>
-          </ion-select>
+          <style>
+            #container {
+              display: grid;
+              gap: 5px;
+            }
+          </style>
+          <div id="container">
+            <ion-select fill="outline" value="apple" class="ion-touched ion-invalid has-focus" label="Favorite Fruit">
+              <ion-select-option value="apple">Apple</ion-select-option>
+            </ion-select>
+            <ion-list>
+              <ion-item>
+                <ion-select fill="outline" value="apple" class="ion-touched ion-invalid has-focus" label="Favorite Fruit">
+                  <ion-select-option value="apple">Apple</ion-select-option>
+                </ion-select>
+              </ion-item>
+            </ion-list>
+          </div>
         `,
           config
         );
 
-        const select = page.locator('ion-select');
-        await expect(select).toHaveScreenshot(screenshot(`select-outline-invalid`));
+        const container = page.locator('#container');
+        await expect(container).toHaveScreenshot(screenshot(`select-outline-invalid`));
       });
       test('should render focused state correctly', async ({ page }) => {
         await page.setContent(
           `
-          <ion-select fill="outline" value="apple" class="ion-focused" label="Favorite Fruit">
-            <ion-select-option value="apple">Apple</ion-select-option>
-          </ion-select>
+          <style>
+            #container {
+              display: grid;
+              gap: 5px;
+            }
+          </style>
+          <div id="container">
+            <ion-select fill="outline" value="apple" class="has-focus" label="Favorite Fruit">
+              <ion-select-option value="apple">Apple</ion-select-option>
+            </ion-select>
+            <ion-list>
+              <ion-item>
+                <ion-select fill="outline" value="apple" class="has-focus" label="Favorite Fruit">
+                  <ion-select-option value="apple">Apple</ion-select-option>
+                </ion-select>
+              </ion-item>
+            </ion-list>
+          </div>
         `,
           config
         );
 
-        const select = page.locator('ion-select');
-        await expect(select).toHaveScreenshot(screenshot(`select-outline-focus`));
+        const container = page.locator('#container');
+        await expect(container).toHaveScreenshot(screenshot(`select-outline-focus`));
       });
       test('should render custom highlight correctly', async ({ page }) => {
         await page.setContent(
@@ -202,15 +365,15 @@ configs({ modes: ['md'], directions: ['ltr'] }).forEach(({ title, screenshot, co
           </style>
 
           <div class="container">
-            <ion-select fill="outline" value="apple" class="custom ion-valid ion-focused" label="Favorite Fruit">
+            <ion-select fill="outline" value="apple" class="custom ion-valid has-focus" label="Favorite Fruit">
               <ion-select-option value="apple">Apple</ion-select-option>
             </ion-select>
 
-            <ion-select fill="outline" value="apple" class="custom ion-touched ion-invalid ion-focused" label="Favorite Fruit">
+            <ion-select fill="outline" value="apple" class="custom ion-touched ion-invalid has-focus" label="Favorite Fruit">
               <ion-select-option value="apple">Apple</ion-select-option>
             </ion-select>
 
-            <ion-select fill="outline" value="apple" class="custom ion-focused" label="Favorite Fruit">
+            <ion-select fill="outline" value="apple" class="custom has-focus" label="Favorite Fruit">
               <ion-select-option value="apple">Apple</ion-select-option>
             </ion-select>
           </div>
@@ -229,39 +392,78 @@ configs({ modes: ['md'], directions: ['ltr'] }).forEach(({ title, screenshot, co
       test('should render bottom highlight', async ({ page }) => {
         await page.setContent(
           `
-          <ion-select label="Label" class="select-expanded"></ion-select>
+          <style>
+            #container {
+              display: grid;
+              gap: 5px;
+            }
+          </style>
+          <div id="container">
+            <ion-select label="Label" class="select-expanded"></ion-select>
+            <ion-list>
+              <ion-item>
+                <ion-select label="Label" class="select-expanded"></ion-select>
+              </ion-item>
+            </ion-list>
+          </div>
         `,
           config
         );
 
-        const select = page.locator('ion-select');
-        await expect(select).toHaveScreenshot(screenshot(`select-no-fill-highlight`));
+        const container = page.locator('#container');
+        await expect(container).toHaveScreenshot(screenshot(`select-no-fill-highlight`));
       });
     });
     test.describe('select: solid', () => {
       test('should render bottom highlight', async ({ page }) => {
         await page.setContent(
           `
-          <ion-select fill="solid" label="Label" class="select-expanded"></ion-select>
+          <style>
+            #container {
+              display: grid;
+              gap: 5px;
+            }
+          </style>
+          <div id="container">
+            <ion-select fill="solid" label="Label" class="select-expanded"></ion-select>
+            <ion-list>
+              <ion-item>
+                <ion-select fill="solid" label="Label" class="select-expanded"></ion-select>
+              </ion-item>
+            </ion-list>
+          </div>
         `,
           config
         );
 
-        const select = page.locator('ion-select');
-        await expect(select).toHaveScreenshot(screenshot(`select-solid-highlight`));
+        const container = page.locator('#container');
+        await expect(container).toHaveScreenshot(screenshot(`select-solid-highlight`));
       });
     });
     test.describe('select: outline', () => {
       test('should render bottom highlight', async ({ page }) => {
         await page.setContent(
           `
-          <ion-select fill="outline" label="Label" class="select-expanded"></ion-select>
+          <style>
+            #container {
+              display: grid;
+              gap: 5px;
+            }
+          </style>
+          <div id="container">
+            <ion-select fill="outline" label="Label" class="select-expanded"></ion-select>
+            <ion-list>
+              <ion-item>
+                <ion-select fill="outline" label="Label" class="select-expanded"></ion-select>
+              </ion-item>
+            </ion-list>
+          </div>
         `,
           config
         );
 
-        const select = page.locator('ion-select');
-        await expect(select).toHaveScreenshot(screenshot(`select-outline-highlight`));
+        const container = page.locator('#container');
+        await expect(container).toHaveScreenshot(screenshot(`select-outline-highlight`));
       });
     });
   });