+++ packages/angular/test/base/src/app/standalone/app-standalone/app.component.ts
@@ -1,12 +1,11 @@
 import { Component } from '@angular/core';
-import { RouterModule } from '@angular/router';
-import { IonRouterOutlet, IonApp } from '@ionic/angular/standalone';
+import { IonApp, IonRouterOutlet } from '@ionic/angular/standalone';
 
 @Component({
   selector: 'app-root-standalone',
   templateUrl: './app.component.html',
   standalone: true,
-  imports: [RouterModule, IonRouterOutlet, IonApp]
+  imports: [IonRouterOutlet, IonApp]
 })
 export class AppComponent {
 }